extracted argument parser
